Add rules doc, move seeds into solvable_seeds folder
Sets up a shared spot for precomputed seed pools in case other games
get this treatment later.
