+++ html-src/rules/klondikealwayssolvable.html
@@ -1,0 +1,20 @@
+<h1>Klondike (Always Solvable)</h1>
+<p>
+Klondike type. 1 deck. Unlimited redeals.
+
+<h3>Object</h3>
+<p>
+Move all cards to the foundations.
+
+<h3>Quick Description</h3>
+<p>
+Just like <a href="klondike.html">Klondike</a>, except new deals are
+drawn from a pool of game numbers that have already been confirmed
+solvable (winnable with <b></b>perfect</b> play), so you should never get stuck
+with an unwinnable deal.
+
+<h3>Notes</h3>
+<p>
+This only applies to fresh deals. If you manually enter a specific game
+number, that exact number is dealt as-is and is not guaranteed to be
+solvable.
